pod: add new WithLabels method
The existing WithLabel method only allows to add one label to a Pod's
definition. If more want to be added the old ones will be overwritten.
This new method accepts a set of labels in the form of a map to define
all the labels required at once.

Also, the isMutationAllowed and the isMountAlreadyInUseInPod methods
haven been modified to return a bool as the name suggests. This has been
used in the different With* methods to leave as early as possible if the
right condition is not satisfied.