BMC package added as a wrapper to some redfish APIs and serial console.
The BMC package allows both redfish access via the gofish library and
the system's ssh-tunneled serial console.

In order to avoid explicit Connect/Disconnect (or Login/Logout) methods
for redfish, every method (e.g. SystemManufacturer) internally connects
before accessing redfish and disconnects right before returning. That
way the session (and the timeout) is automatically created for that
access only, minimizing the risk of leaving it open.

Every method for redfish access needs a system index. Usually, there's a
1:1 relationship between BMC:Server but it might not be always the case,
so I believe it's a good idea to allow the user to explicitly set the
index, even when most of the test cases will use "0" for it.

The SSH timeout is only used for opening the ssh tunnel that is needed
for the bmc ssh'd cli to call the command that opens the serial console,
although it might be used in the future for direct ssh'd cli access to
the bmc.

The method OpenSerialConsole accepts a custom cli command to be sent to
the bmc's ssh'd cli. If no command is provided, it will try to use the
right one based on the first system's manufacturer as a best effort.