BMC: Added CLI commands runner and Serial Console access.
BMC's CLI commands can be run using the bmc.RunCLICommand method,
which connects to the CLI via SSH. The method blocks until the command
ends or a timeout happens, whatever comes first. The command's output is
copied into the stdout & stderr string vars. The underlying SSH session
is created and closed internally after the command finishes.

The new method OpenSerialConsole returns a piped reader and writer
interfaces that can be used to interactively receive/send commands/text
to/from the serial console. Since the serial console is tunneled in a
SSH session, the CloseSerialConsole method must be called to release
that session to prevent leaks that would prevent CLI commands to fail
due to maximum concurrent SSH sessions reached in the BMC.

Also, added SSH port as parameter for BMC's New() method