bmc: add power usage and more reset actions (#425)
Adds the ability to collect power usage from the Redfish client. To support that, an index into the power control list is added and a supporting method to set it.

Also added a method to perform any reset action as well as a few other common reset actions such as power on/off/cycle.