bmc: refactor creation and configuration
This commit brings the creation and configuration for the BMC client more in line with the rest of the packages and allows for Redfish and SSH access to be configured separately.

Depending on the use case, the Redfish user or the SSH user must be provided, with a few other options available for configuration. A summary of these configuration methods is provided:

```
func New(host string) *BMC

Redfish access only:
  Mandatory:
    func (bmc *BMC) WithRedfishUser(user, pass string) *BMC
  Optional:
    func (bmc *BMC) WithRedfishTimeout(timeout time.Duration) *BMC
    func (bmc *BMC) WithRedfishSystemIndex(index int) *BMC
    func (bmc *BMC) WithRedfishPowerControlIndex(index int) *BMC

SSH Access (cli commands) only:
  Mandatory:
    func (bmc *BMC) WithSSHUser(user, pass string) *BMC
  Optional:
    func (bmc *BMC) WithSSHPort(port uint16) *BMC
    func (bmc *BMC) WithSSHTimeout(timeout time.Duratin) *BMC
```

Additionally, the unit tests have been updated to match these new public methods and all use assert.