sync: refactor sync process
Previously, the sync process would attempt to reverse the changes to be applied to the local directory and diff it against the cloned directory. If they were the same, the local directory would be reverted using git. If they were different, the cloned directory would have the opposite changes applied and then be copied to the local directory.

This relied on the local directory already being in git and would cause errors when it was not tracked in git. This PR changes this process to apply the desired updates to the cloned directory, diff it against the local, and overwrite the local if there is a difference. This does not rely on git to revert the changes and should halve the number of changes to files.