clients: add extra log on error and use glog (#761)
It turns out there is a case in the clients package where the New function will return nil due to an error but not log the error anywhere. This PR adds an additional log for this code path as well as switches the log statements to use glog instead of the standard library's log package. glog is used elsewhere in the package so this makes it consistent with the package and the rest of the repo.