ocm: add WaitForAllPoliciesComplianceState function
Although there is already a way to wait for an individual policy to reach a certain compliance state, there is no way to wait for all policies to reach a compliance state. This PR addresses this gap by adding the WaitForAllPoliciesComplianceState function.

In the context of a spoke cluster, this function can be used directly to wait for policies to all go compliant. For a hub cluster, one may provide ListOptions to specify a label or namespace to wait on the policies of a given ManagedCluster.