sync: update label job if condition
By default, GitHub actions will apply a `success()` condition to the if unless there is another status check function used. Since the label function is meant to run even if previous steps failed, we need to use a status check function to replace the default success condition. This PR adds a `!cancelled()` condition to address that and prevent issues where the job runs even if dependencies have been cancelled.