mco: update machine config api dependency (#892)
Previously, the machine config types were taken from the operator repo but they have since been moved to the openshift/api repo. This PR switches the mco package to use the new repo for the types.

Unfortunately, a few other operators depend on the old versions of the types so it remains as an indirect dependency.