ci: fix sync-libs testing
Although sync-libs uses the sync branch for its changes, all of the subsequent tests would run on the default branch. This hides the errors from the tests unless main also has broken CI.

This PR updates all the workflows that sync-libs calls to take a branch input that determines what to checkout. Other triggers for the workflows just use the GITHUB_SHA for that event type.