nto: update dependency and import consistency
Switched all the v2 imports to performanceprofilev2 to improve import consistency. The dependency on cluster-node-tuning-operator has been increased to the latest version that does not require us to update the k8s libraries to 1.32.z. This is for supporting aarch64 clusters that require the kernel page size field to be present otherwise the hugepages configuration will be rejected by the admission webhooks.