PoC: common get/validate for builder
I decided I'd try an attempt at a PoC, it draws a lot from Trey and Brandon's PoCs. Like Trey's it assumes that the controller-runtime client is the only one we'll use. I went with generics to allow us to return concrete types, but it has the weird workaround due to go's generics and getting the type without the pointer but still enforcing the interface on the pointer. Rust allows something like `fn myFunc<T>(param: T) where for<'a> &'a T: MyInterface` but unfortunately this is not possible in a single type parameter in Go.