updated to add example for embedding
I really liked the embedding idea from Trey so I've made an attempt at creating a struct that could be embedded and combine with the Builder interface to offer common methods. The generics get pretty ugly for it, but it does work and pass all the tests.

This also helps to deal with the good point that Trey raised that without embedding we'd be artifically lowering the test coverage.

We'd still run into issues surrounding methods like Update, but I imagine these could be handled similar to my original solution (and what I've used for NewBuilder and Pull in this commit) where we just create a new method that doesn't go on the embedded struct and then use the common function.

We may also be able to use different embedded structs, but I think the amount of boilerplate from that would mean we'd embed a common struct and then implement methods on the more specific structs. Either would work, but I haven't looked into which one ultimately seems better.