oran: add WaitForPhaseAfter method (#1013)
Currently, there is no way to assert that a ProvisioningRequest only becomes fulfilled after a change has been made. This is especially impactful with the use of the O2IMS API, where the information about specific policies is not returned. Without being able to check the policies to know if the ProvisioningRequest has implemented any day2 changes, it becomes necessary to know the ProvisioningRequest had not already been fulfilled at a certain time.

Conveniently, the existing WaitUntilFulfilled method has been redefined in terms of WaitForPhaseAfter by allowing zero times to turn off the check for update times.

This situation also applies to other phases than fulfilled so the method is now more general and can support all the ways to wait for the ProvisioningPhase.