ptp: add PtpOperatorConfig resource and unit tests  (#1022)
* sync: add types for ptp

PTP types have a dependency on openshift/api that causes issues with multiple other dependencies. However, the ptp-operator api package is easy to sync so thes PR adds a new sync config for it.

* ptp: add PtpConfig type and unit tests

* Added PtpConfig type from the ptp-operator.
* Added GetE810Plugin/WithE810Plugin methods for accessing the Intel-specific plugin data when available, using types from the linuxptp-daemon repository.
* Added unit tests for all functions.

* ptp: add PtpOperatorConfig resource and unit tests

Added the PtpOperatorConfig resource to the ptp package along with corresponding unit tests. Since this is a singleton resource, the name and namespace are constants and there is no option to create or delete.